moved ThemeContext.tsx from context folder to provider folder and renamed the file to ThemeProvider.tsx
