+++ src/components/ThemeToggle.tsx
@@ -1,5 +1,5 @@
 import React from "react"
-import { useTheme } from "../contexts/ThemeContext"
+import { useTheme } from "../providers/ThemeProvider"
 import { Moon, Sun } from "lucide-react"
 import "../css/ThemeToggle.css"
